An explicit runtime storage module
